Add docsy module to Hugo
We use the docsy module for this documentation site. Docsy is a Hugo
module specifically designed for software project websites and their
documentation.

Specifically, we opted for installing docsy as module and to create
the site from scratch, rather than basing it on the example template
provided by upstream.

See:
- https://www.docsy.dev/docs/get-started/docsy-as-module/start-from-scratch/

Signed-off-by: Wolodja Wentland <wwentland@wavecon.de>